+++ core/src/lib/scan.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2000-2011 Free Software Foundation Europe e.V.
    Copyright (C) 2016-2016 Planets Communications B.V.
-   Copyright (C) 2016-2019 Bareos GmbH & Co. KG
+   Copyright (C) 2016-2020 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -202,7 +202,7 @@ char* next_arg(char** s)
  *  argk[2] = arg3
  *  argv[2] =
  */
-int ParseArgs(POOLMEM* cmd,
+int ParseArgs(const POOLMEM* cmd,
               POOLMEM*& args,
               int* argc,
               char** argk,
@@ -246,7 +246,7 @@ int ParseArgs(POOLMEM* cmd,
  *  argk[2] = arg3
  *  argv[2] =
  */
-int ParseArgsOnly(POOLMEM* cmd,
+int ParseArgsOnly(const POOLMEM* cmd,
                   POOLMEM*& args,
                   int* argc,
                   char** argk,